Refactor download files
